migrate CPE generation logic to dedicated package
