+++ syft/pkg/cataloger/common/cpe/field_candidate_filter.go
@@ -1,0 +1,11 @@
+package cpe
+
+type filterFieldCandidateFn func(fieldCandidate) bool
+
+func filterFieldCandidatesBySubselection(c fieldCandidate) bool {
+	return c.disallowSubSelections
+}
+
+func filterFieldCandidatesByDelimiterVariations(c fieldCandidate) bool {
+	return c.disallowDelimiterVariations
+}
