+++ syft/pkg/cataloger/common/cpe/generate_test.go
@@ -1,4 +1,4 @@
-package cataloger
+package cpe
 
 import (
 	"fmt"
@@ -400,7 +400,7 @@ func TestGeneratePackageCPEs(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			actual := generatePackageCPEs(test.p)
+			actual := Generate(test.p)
 
 			expectedCpeSet := set.NewStringSet(test.expected...)
 			actualCpeSet := set.NewStringSet()
@@ -497,136 +497,6 @@ func TestCandidateProducts(t *testing.T) {
 	}
 }
 
-func TestCandidateProductForGo(t *testing.T) {
-	tests := []struct {
-		pkg      string
-		expected string
-	}{
-		{
-			pkg:      "github.com/someone/something",
-			expected: "something",
-		},
-		{
-			pkg:      "golang.org/x/xerrors",
-			expected: "x/xerrors",
-		},
-		{
-			pkg:      "gopkg.in/yaml.v2",
-			expected: "yaml.v2",
-		},
-		{
-			pkg:      "place",
-			expected: "",
-		},
-		{
-			pkg:      "place.com/",
-			expected: "",
-		},
-		{
-			pkg:      "place.com/someone-or-thing",
-			expected: "",
-		},
-		{
-			pkg:      "google.golang.org/genproto/googleapis/rpc/status",
-			expected: "genproto",
-		},
-		{
-			pkg:      "github.com/someone/something/long/package/name",
-			expected: "something",
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.pkg, func(t *testing.T) {
-			assert.Equal(t, test.expected, candidateProductForGo(test.pkg))
-		})
-	}
-}
-
-func TestCandidateVendorForGo(t *testing.T) {
-	tests := []struct {
-		pkg      string
-		expected string
-	}{
-		{
-			pkg:      "github.com/someone/something",
-			expected: "someone",
-		},
-		{
-			pkg:      "golang.org/x/xerrors",
-			expected: "golang",
-		},
-		{
-			pkg:      "gopkg.in/yaml.v2",
-			expected: "",
-		},
-		{
-			pkg:      "place",
-			expected: "",
-		},
-		{
-			pkg:      "place.com/",
-			expected: "",
-		},
-		{
-			pkg:      "place.com/someone-or-thing",
-			expected: "",
-		},
-		{
-			pkg:      "google.golang.org/genproto/googleapis/rpc/status",
-			expected: "google",
-		},
-		{
-			pkg:      "github.com/someone/something/long/package/name",
-			expected: "someone",
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.pkg, func(t *testing.T) {
-			assert.Equal(t, test.expected, candidateVendorForGo(test.pkg))
-		})
-	}
-}
-
-func Test_normalizeAuthorName(t *testing.T) {
-	tests := []struct {
-		input   string
-		expects string
-	}{
-		{
-			// note: extra spaces
-			input:   "  Alex Goodman  ",
-			expects: "alex_goodman",
-		},
-		{
-			input:   "Alex Goodman",
-			expects: "alex_goodman",
-		},
-		{
-			input:   "Alex.Goodman",
-			expects: "alex_goodman",
-		},
-		{
-			input:   "Alex.Goodman",
-			expects: "alex_goodman",
-		},
-		{
-			input:   "AlexGoodman",
-			expects: "alexgoodman",
-		},
-		{
-			input:   "The Apache Software Foundation",
-			expects: "apache_software_foundation",
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.input, func(t *testing.T) {
-			assert.Equal(t, test.expects, normalizeName(test.input))
-		})
-	}
-}
-
 func Test_generateSubSelections(t *testing.T) {
 	tests := []struct {
 		field    string
@@ -703,449 +573,6 @@ func Test_addSeparatorVariations(t *testing.T) {
 			val := newCPRFieldCandidateSet(test.input...)
 			addDelimiterVariations(val)
 			assert.ElementsMatch(t, test.expected, val.values())
-		})
-	}
-}
-
-func Test_productsFromArtifactAndGroupIDs(t *testing.T) {
-	tests := []struct {
-		groupIDs   []string
-		artifactID string
-		expected   []string
-	}{
-		{
-			groupIDs:   []string{"org.sonatype.nexus"},
-			artifactID: "nexus-extender",
-			expected:   []string{"nexus", "nexus-extender"},
-		},
-		{
-			groupIDs: []string{"org.sonatype.nexus"},
-			expected: []string{"nexus"},
-		},
-		{
-			groupIDs:   []string{"org.jenkins-ci.plugins"},
-			artifactID: "ant",
-			expected:   []string{"ant"},
-		},
-		{
-			groupIDs:   []string{"org.jenkins-ci.plugins"},
-			artifactID: "antisamy-markup-formatter",
-			expected:   []string{"antisamy-markup-formatter"},
-		},
-		{
-			groupIDs:   []string{"io.jenkins.plugins"},
-			artifactID: "aws-global-configuration",
-			expected:   []string{"aws-global-configuration"},
-		},
-		{
-			groupIDs:   []string{"com.cloudbees.jenkins.plugins"},
-			artifactID: "cloudbees-servicenow-jenkins-plugin",
-			expected:   []string{"cloudbees-servicenow-jenkins-plugin"},
-		},
-		{
-			groupIDs:   []string{"com.atlassian.confluence.plugins"},
-			artifactID: "confluence-mobile-plugin",
-			expected:   []string{"confluence-mobile-plugin"},
-		},
-		{
-			groupIDs:   []string{"com.atlassian.confluence.plugins"},
-			artifactID: "confluence-view-file-macro",
-			expected:   []string{"confluence-view-file-macro"},
-		},
-		{
-			groupIDs:   []string{"com.google.guava"},
-			artifactID: "failureaccess",
-			expected:   []string{"failureaccess"},
-		},
-	}
-	for _, test := range tests {
-		t.Run(strings.Join(test.groupIDs, ",")+":"+test.artifactID, func(t *testing.T) {
-			actual := productsFromArtifactAndGroupIDs(test.artifactID, test.groupIDs)
-			assert.ElementsMatch(t, test.expected, actual, "different products")
-		})
-	}
-}
-
-func Test_candidateProductsForJava(t *testing.T) {
-	tests := []struct {
-		name     string
-		pkg      pkg.Package
-		expected []string
-	}{
-		{
-			name: "duplicate groupID in artifactID field",
-			pkg: pkg.Package{
-				Metadata: pkg.JavaMetadata{
-					PomProperties: &pkg.PomProperties{
-						GroupID:    "org.sonatype.nexus",
-						ArtifactID: "org.sonatype.nexus",
-					},
-				},
-			},
-			expected: []string{"nexus"},
-		},
-		{
-			name: "detect groupID-like value in artifactID field",
-			pkg: pkg.Package{
-				Metadata: pkg.JavaMetadata{
-					PomProperties: &pkg.PomProperties{
-						ArtifactID: "org.sonatype.nexus",
-					},
-				},
-			},
-			expected: []string{"nexus"},
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			actual := candidateProductsForJava(test.pkg)
-			assert.ElementsMatch(t, test.expected, actual, "different products")
-		})
-	}
-}
-
-func Test_vendorsFromGroupIDs(t *testing.T) {
-	tests := []struct {
-		groupIDs []string
-		expected []string
-	}{
-		{
-			groupIDs: []string{"org.sonatype.nexus"},
-			expected: []string{"sonatype", "nexus"},
-		},
-		{
-			groupIDs: []string{"org.sonatype.nexus"},
-			expected: []string{"sonatype", "nexus"},
-		},
-		{
-			groupIDs: []string{"org.sonatype.nexus"},
-			expected: []string{"sonatype", "nexus"},
-		},
-		{
-			groupIDs: []string{"org.jenkins-ci.plugins"},
-			expected: []string{"jenkins-ci", "jenkins"},
-		},
-		{
-			groupIDs: []string{"org.jenkins-ci.plugins"},
-			expected: []string{"jenkins-ci", "jenkins"},
-		},
-		{
-			groupIDs: []string{"io.jenkins.plugins"},
-			expected: []string{"jenkins"},
-		},
-		{
-			groupIDs: []string{"com.cloudbees.jenkins.plugins"},
-			expected: []string{"cloudbees", "jenkins"},
-		},
-		{
-			groupIDs: []string{"com.atlassian.confluence.plugins"},
-			expected: []string{"atlassian", "confluence"},
-		},
-		{
-			groupIDs: []string{"com.atlassian.confluence.plugins"},
-			expected: []string{"atlassian", "confluence"},
-		},
-		{
-			groupIDs: []string{"com.google.guava"},
-			expected: []string{"google", "guava"},
-		},
-	}
-	for _, test := range tests {
-		t.Run(strings.Join(test.groupIDs, ","), func(t *testing.T) {
-			assert.ElementsMatch(t, test.expected, vendorsFromGroupIDs(test.groupIDs).values(), "different vendors")
-		})
-	}
-}
-
-func Test_groupIDsFromJavaPackage(t *testing.T) {
-	tests := []struct {
-		name    string
-		pkg     pkg.Package
-		expects []string
-	}{
-		{
-			name: "go case",
-			pkg: pkg.Package{
-				Metadata: pkg.JavaMetadata{
-					PomProperties: &pkg.PomProperties{
-						GroupID: "io.jenkins-ci.plugin.thing",
-					},
-				},
-			},
-			expects: []string{"io.jenkins-ci.plugin.thing"},
-		},
-		{
-			name: "from artifactID",
-			pkg: pkg.Package{
-				Metadata: pkg.JavaMetadata{
-					PomProperties: &pkg.PomProperties{
-						ArtifactID: "io.jenkins-ci.plugin.thing",
-					},
-				},
-			},
-			expects: []string{"io.jenkins-ci.plugin.thing"},
-		},
-		{
-			name: "from main Extension-Name field",
-			pkg: pkg.Package{
-				Metadata: pkg.JavaMetadata{
-					Manifest: &pkg.JavaManifest{
-						Main: map[string]string{
-							"Extension-Name": "io.jenkins-ci.plugin.thing",
-						},
-					},
-				},
-			},
-			expects: []string{"io.jenkins-ci.plugin.thing"},
-		},
-		{
-			name: "from named section Extension-Name field",
-			pkg: pkg.Package{
-				Metadata: pkg.JavaMetadata{
-					Manifest: &pkg.JavaManifest{
-						NamedSections: map[string]map[string]string{
-							"section": {
-								"Extension-Name": "io.jenkins-ci.plugin.thing",
-							},
-						},
-					},
-				},
-			},
-			expects: []string{"io.jenkins-ci.plugin.thing"},
-		},
-		{
-			name: "from main field",
-			pkg: pkg.Package{
-				Metadata: pkg.JavaMetadata{
-					Manifest: &pkg.JavaManifest{
-						Main: map[string]string{
-							// positive cases
-							"Automatic-Module-Name":    "io.jenkins-ci.plugin.1",
-							"Extension-Name":           "io.jenkins-ci.plugin.2",
-							"Specification-Vendor":     "io.jenkins-ci.plugin.3",
-							"Implementation-Vendor":    "io.jenkins-ci.plugin.4",
-							"Bundle-SymbolicName":      "io.jenkins-ci.plugin.5",
-							"Implementation-Vendor-Id": "io.jenkins-ci.plugin.6",
-							"Package":                  "io.jenkins-ci.plugin.7",
-							"Implementation-Title":     "io.jenkins-ci.plugin.8",
-							"Main-Class":               "io.jenkins-ci.plugin.9",
-							"Bundle-Activator":         "io.jenkins-ci.plugin.10",
-						},
-					},
-				},
-			},
-			expects: []string{
-				"io.jenkins-ci.plugin.1",
-				"io.jenkins-ci.plugin.2",
-				"io.jenkins-ci.plugin.3",
-				"io.jenkins-ci.plugin.4",
-				"io.jenkins-ci.plugin.5",
-				"io.jenkins-ci.plugin.6",
-				"io.jenkins-ci.plugin.7",
-				"io.jenkins-ci.plugin.8",
-				"io.jenkins-ci.plugin.9",
-				"io.jenkins-ci.plugin.10",
-			},
-		},
-		{
-			name: "from main field - negative cases",
-			pkg: pkg.Package{
-				Metadata: pkg.JavaMetadata{
-					Manifest: &pkg.JavaManifest{
-						Main: map[string]string{
-							// negative cases
-							"Extension-Name": "not.a-group.id",
-							"bogus":          "io.jenkins-ci.plugin.please-dont-find-me",
-						},
-					},
-				},
-			},
-			expects: nil,
-		},
-		{
-			name: "from named section field",
-			pkg: pkg.Package{
-				Metadata: pkg.JavaMetadata{
-					Manifest: &pkg.JavaManifest{
-						NamedSections: map[string]map[string]string{
-							"section": {
-								// positive cases
-								"Automatic-Module-Name":    "io.jenkins-ci.plugin.1",
-								"Extension-Name":           "io.jenkins-ci.plugin.2",
-								"Specification-Vendor":     "io.jenkins-ci.plugin.3",
-								"Implementation-Vendor":    "io.jenkins-ci.plugin.4",
-								"Bundle-SymbolicName":      "io.jenkins-ci.plugin.5",
-								"Implementation-Vendor-Id": "io.jenkins-ci.plugin.6",
-								"Package":                  "io.jenkins-ci.plugin.7",
-								"Implementation-Title":     "io.jenkins-ci.plugin.8",
-								"Main-Class":               "io.jenkins-ci.plugin.9",
-								"Bundle-Activator":         "io.jenkins-ci.plugin.10",
-							},
-						},
-					},
-				},
-			},
-			expects: []string{
-				"io.jenkins-ci.plugin.1",
-				"io.jenkins-ci.plugin.2",
-				"io.jenkins-ci.plugin.3",
-				"io.jenkins-ci.plugin.4",
-				"io.jenkins-ci.plugin.5",
-				"io.jenkins-ci.plugin.6",
-				"io.jenkins-ci.plugin.7",
-				"io.jenkins-ci.plugin.8",
-				"io.jenkins-ci.plugin.9",
-				"io.jenkins-ci.plugin.10",
-			},
-		},
-		{
-			name: "from named section field - negative cases",
-			pkg: pkg.Package{
-				Metadata: pkg.JavaMetadata{
-					Manifest: &pkg.JavaManifest{
-						NamedSections: map[string]map[string]string{
-							"section": {
-								// negative cases
-								"Extension-Name": "not.a-group.id",
-								"bogus":          "io.jenkins-ci.plugin.please-dont-find-me",
-							},
-						},
-					},
-				},
-			},
-			expects: nil,
-		},
-		{
-			name: "no manifest or pom info",
-			pkg: pkg.Package{
-				Metadata: pkg.JavaMetadata{},
-			},
-			expects: nil,
-		},
-		{
-			name:    "no java info",
-			pkg:     pkg.Package{},
-			expects: nil,
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			assert.ElementsMatch(t, test.expects, groupIDsFromJavaPackage(test.pkg))
-		})
-	}
-}
-
-func Test_artifactIDFromJavaPackage(t *testing.T) {
-	tests := []struct {
-		name    string
-		pkg     pkg.Package
-		expects string
-	}{
-		{
-			name: "go case",
-			pkg: pkg.Package{
-				Metadata: pkg.JavaMetadata{
-					PomProperties: &pkg.PomProperties{
-						ArtifactID: "cloudbees-installation-manager",
-					},
-				},
-			},
-			expects: "cloudbees-installation-manager",
-		},
-		{
-			name: "ignore groupID-like things",
-			pkg: pkg.Package{
-				Metadata: pkg.JavaMetadata{
-					PomProperties: &pkg.PomProperties{
-						ArtifactID: "io.jenkins-ci.plugin.thing",
-					},
-				},
-			},
-			expects: "",
-		},
-		{
-			name:    "no java info",
-			pkg:     pkg.Package{},
-			expects: "",
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			assert.Equal(t, test.expects, artifactIDFromJavaPackage(test.pkg))
-		})
-	}
-}
-
-func Test_vendorsFromJavaManifestNames(t *testing.T) {
-	tests := []struct {
-		name    string
-		pkg     pkg.Package
-		expects []string
-	}{
-		{
-			name: "from manifest named section fields",
-			pkg: pkg.Package{
-				Metadata: pkg.JavaMetadata{
-					Manifest: &pkg.JavaManifest{
-						NamedSections: map[string]map[string]string{
-							"section": {
-								// positive cases
-								"Specification-Vendor":  "Alex Goodman",
-								"Implementation-Vendor": "William Goodman",
-							},
-						},
-					},
-				},
-			},
-			expects: []string{"alex_goodman", "william_goodman"},
-		},
-		{
-			name: "from manifest named section fields - negative cases",
-			pkg: pkg.Package{
-				Metadata: pkg.JavaMetadata{
-					Manifest: &pkg.JavaManifest{
-						NamedSections: map[string]map[string]string{
-							"section": {
-								// negative cases
-								"Specification-Vendor":     "io.jenkins-ci.plugin.thing",
-								"Implementation-Vendor-ID": "William Goodman",
-							},
-						},
-					},
-				},
-			},
-			expects: nil,
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			assert.ElementsMatch(t, test.expects, vendorsFromJavaManifestNames(test.pkg).values())
-		})
-	}
-}
-
-func Test_normalizeTitle(t *testing.T) {
-	tests := []struct {
-		input   string
-		expects string
-	}{
-		{
-			// note: extra spaces
-			input:   "  Alex Goodman  ",
-			expects: "alexgoodman",
-		},
-		{
-			input:   "Alex Goodman, LLC",
-			expects: "alexgoodman",
-		},
-		{
-			input:   "alex.goodman",
-			expects: "alex.goodman",
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.input, func(t *testing.T) {
-			assert.Equal(t, test.expects, normalizeTitle(test.input))
 		})
 	}
 }
